requirements.txt
pyserial==3.5
torch==2.4.1
tqdm==4.66.2
